Disable touch-screen auto-detection
When touch screen is detected, XCSoar renders larger buttons and other
screen controls. Larger control size is more preferable for Openvario
even on machines with no touch screen is installed: it is easier to
click larger controls with the stick control and labels actually fit the
buttons with the larger font sizes.