Including new adapter board DS2
The new adapter board DS2 is a useful further development of the original adapter board. The creation of new machines is necessary because all four of the ttyS interfaces can now be used. In addition, they are new and sensibly arranged. All other changes do not require any software adjustments.