Timeout usbstick automount after 2 seconds
When user tries to perform any operation that involves USB stick (e.g.
"Download IGC") and USB stick is not inserted, openvario appears to
"freeze" - becomes unresponsive, with no way to recover.

This is because it waits for /dev/sda1 device to appear in order to
mount it.  With the default 90s timeout the system appears to be stuck.

This lowers the timeout to 2 seconds to "unstuck" the system quickly
enough before user gets frustrated.