Further increase minimum CPU voltage to 1.2V
This change, when applied, changes the OPP voltages for both 144MHz and 312MHz
from 1V to 1.2V when creating Openvario images. This fixes freezes previously
encountered. It is a "follow up" change to PR #324. It turned out that even
at 1.1V OV occasionally crashed.

For more information on the why and how of this change please see comment in
PR #324