Extensions for system info functionality to script system-info.sh.
The system info script gives an overview which packages and which versions are installed.
It also shows IP addresses, kernel, host, system daemon status and other info.

A bunch of new system information was added:
- KERNEL_VERSION
- MAC address of eth0 (also relevant for backups)
- HOSTNAME
- versions of:
  - I2C_TOOLS
  - E2FSPROGS
  - USB_MODESWITCH
- daemon status of:
  - SSH
  - variod
  - sensord

The output is no more done with a "dialog" message box but directly echoed to the console.

Yes, querying the version of the daemons by a "opkg list-installed" is not as portable as desired:
https://github.com/Openvario/meta-openvario/pull/320#discussion_r952617510
But this is an inherited flaw from earlier system-info versions.
It should be improved separately from this commit.